requirements.txt
annotated-types==0.7.0
bidict==0.23.1
blinker==1.9.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
dnspython==2.7.0
eventlet==0.39.0
filelock==3.17.0
Flask==3.1.0
Flask-Cors==5.0.0
Flask-SocketIO==5.5.1
fonttools==4.61.1
fsspec==2024.12.0
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
joblib==1.4.2
kiwisolver==1.4.9
MarkupSafe==3.0.2
matplotlib==3.10.8
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.2
packaging==24.2
pandas==2.2.3
pillow==11.1.0
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.1
pymongo==4.11
pyparsing==3.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.11.2
python-socketio==5.12.1
pytz==2024.2
requests==2.32.3
scikit-learn==1.6.0
scipy==1.15.1
setuptools==80.9.0
simple-websocket==1.1.0
six==1.17.0
sympy==1.13.1
threadpoolctl==3.5.0
torch==2.6.0
torchaudio==2.6.0
torchvision==0.21.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
Werkzeug==3.1.3
wsproto==1.2.0
